requirements.txt
aiosqlite==0.21.0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2